+++ service/src/main/java/org/ehrbase/aql/definition/FunctionDefinition.java
@@ -106,4 +106,9 @@ public class FunctionDefinition implements I_VariableDefinition {
     public void setHidden(boolean hidden) {
 
     }
+
+    @Override
+    public void setAlias(String alias) {
+
+    }
 }