+++ service/src/main/java/org/ehrbase/dao/access/interfaces/I_EhrAccess.java
@@ -22,6 +22,8 @@
 package org.ehrbase.dao.access.interfaces;
 
 import com.nedap.archie.rm.datastructures.ItemStructure;
+import com.nedap.archie.rm.datavalues.DvCodedText;
+import com.nedap.archie.rm.datavalues.DvText;
 import org.ehrbase.api.exception.InternalServerException;
 import org.ehrbase.api.exception.InvalidApiParameterException;
 import org.ehrbase.dao.access.jooq.EhrAccess;
@@ -29,7 +31,6 @@ import org.ehrbase.dao.access.util.ContributionDef;
 import org.ehrbase.jooq.pg.tables.records.EhrRecord;
 import com.nedap.archie.rm.ehr.EhrStatus;
 
-import java.sql.Timestamp;
 import java.util.Map;
 import java.util.UUID;
 
@@ -124,7 +125,7 @@ public interface I_EhrAccess extends I_SimpleCRUD {
     }
 
     /**
-     * retrieve the Ehr entry from its id
+     * Retrieve the Ehr entry from its ID (incl latest STATUS).
      *
      * @param domainAccess SQL access
      * @param ehrId        the Ehr UUID
@@ -161,6 +162,14 @@ public interface I_EhrAccess extends I_SimpleCRUD {
     }
 
     void setModifiable(Boolean modifiable);
+
+    void setArchetypeNodeId(String archetypeNodeId);
+
+    String getArchetypeNodeId();
+
+    void setName(DvText name);
+
+    void setName(DvCodedText name);
 
     void setQueryable(Boolean queryable);
 
@@ -252,32 +261,4 @@ public interface I_EhrAccess extends I_SimpleCRUD {
      * @return Latest EHR_STATUS
      */
     EhrStatus getStatus();
-
-    /**
-     * Get latest version number of EHR_STATUS by versioned object UID.
-     * @param domainAccess access
-     * @param ehrStatusId versioned object UID
-     * @return version number
-     */
-    Integer getLastVersionNumberOfStatus(I_DomainAccess domainAccess, UUID ehrStatusId);
-
-    /**
-     * Get a specific version number of the associated EHR_STATUS of this EhrAccess by timestamp.
-     * General idea behind the algorithm: 'what version was the top version at moment T?'
-     * @param time Timestamp
-     * @return version number
-     */
-    int getEhrStatusVersionFromTimeStamp(Timestamp time);
-
-    /**
-     * Get initial time (or time of oldest record) of the status object linked to the EHR of this EhrAccess instance.
-     * @return time as {@link Timestamp}
-     */
-    Timestamp getInitialTimeOfVersionedEhrStatus();
-
-    /**
-     * Get number of available versions of EhrStatus' linked to this EhrAccess instance.
-     * @return Number of versions
-     */
-    Integer getNumberOfEhrStatusVersions();
 }