+++ service/src/main/java/org/ehrbase/service/EhrServiceImp.java
@@ -33,10 +33,9 @@ import org.ehrbase.api.definitions.ServerConfig;
 import org.ehrbase.api.definitions.StructuredString;
 import org.ehrbase.api.definitions.StructuredStringFormat;
 import org.ehrbase.api.dto.EhrStatusDto;
-import org.ehrbase.api.exception.InternalServerException;
-import org.ehrbase.api.exception.ObjectNotFoundException;
-import org.ehrbase.api.exception.StateConflictException;
+import org.ehrbase.api.exception.*;
 import org.ehrbase.api.service.EhrService;
+import org.ehrbase.api.service.ValidationService;
 import org.ehrbase.dao.access.interfaces.*;
 import org.ehrbase.dao.access.jooq.AttestationAccess;
 import org.ehrbase.serialisation.CanonicalJson;
@@ -58,21 +57,34 @@ import java.util.UUID;
 @Service
 @Transactional()
 public class EhrServiceImp extends BaseService implements EhrService {
-    public static final String MODIFIABLE = "modifiable";
-    public static final String QUERYABLE = "queryable";
-    public static final String SUBJECT_ID = "subjectId";
-    public static final String SUBJECT_NAMESPACE = "subjectNamespace";
     public static final String DESCRIPTION = "description";
     private Logger logger = LoggerFactory.getLogger(getClass());
+    private final ValidationService validationService;
 
     @Autowired
-    public EhrServiceImp(KnowledgeCacheService knowledgeCacheService, DSLContext context, ServerConfig serverConfig) {
-
+    public EhrServiceImp(KnowledgeCacheService knowledgeCacheService, ValidationService validationService, DSLContext context, ServerConfig serverConfig) {
         super(knowledgeCacheService, context, serverConfig);
+        this.validationService = validationService;
     }
 
     @Override
     public UUID create(EhrStatus status, UUID ehrId) {
+
+        try {
+            validationService.check(status);
+        } catch (Exception e) {
+            // rethrow if this class, but wrap all others in InternalServerException
+            if (e.getClass().equals(UnprocessableEntityException.class))
+                throw (UnprocessableEntityException) e;
+            if (e.getClass().equals(IllegalArgumentException.class))
+                throw new ValidationException(e);
+            if (e.getClass().equals(ValidationException.class))
+                throw e;
+            else if (e.getClass().equals(org.ehrbase.validation.constraints.wrappers.ValidationException.class))
+                throw new ValidationException(e);
+            else
+                throw new InternalServerException(e);
+        }
 
         if (status == null) {   // in case of new status with default values
             status = new EhrStatus();
@@ -185,6 +197,23 @@ public class EhrServiceImp extends BaseService implements EhrService {
 
     @Override
     public Optional<EhrStatus> updateStatus(UUID ehrId, EhrStatus status) {
+
+        try {
+            validationService.check(status);
+        } catch (Exception e) {
+            // rethrow if this class, but wrap all others in InternalServerException
+            if (e.getClass().equals(UnprocessableEntityException.class))
+                throw (UnprocessableEntityException) e;
+            if (e.getClass().equals(IllegalArgumentException.class))
+                throw new ValidationException(e);
+            if (e.getClass().equals(ValidationException.class))
+                throw e;
+            else if (e.getClass().equals(org.ehrbase.validation.constraints.wrappers.ValidationException.class))
+                throw new ValidationException(e);
+            else
+                throw new InternalServerException(e);
+        }
+
         //pre-step: check for valid ehrId
         if (hasEhr(ehrId).equals(Boolean.FALSE)) {
             throw new ObjectNotFoundException("ehr", "No EHR found with given ID: " + ehrId.toString());
@@ -249,7 +278,7 @@ public class EhrServiceImp extends BaseService implements EhrService {
     @Override
     public Integer getEhrStatusVersionByTimestamp(UUID ehrUid, Timestamp timestamp) {
         I_EhrAccess ehrAccess = I_EhrAccess.retrieveInstance(getDataAccess(), ehrUid);
-        return ehrAccess.getEhrStatusVersionFromTimeStamp(timestamp);
+        return ehrAccess.getStatusAccess().getEhrStatusVersionFromTimeStamp(timestamp);
     }
 
     /**
@@ -261,7 +290,7 @@ public class EhrServiceImp extends BaseService implements EhrService {
         try {
             I_EhrAccess ehrAccess = I_EhrAccess.retrieveInstance(getDataAccess(), ehrStatusId);
             UUID statusId = ehrAccess.getStatusId();
-            Integer version = ehrAccess.getLastVersionNumberOfStatus(getDataAccess(), statusId);
+            Integer version = I_StatusAccess.getLatestVersionNumber(getDataAccess(), statusId);
 
             return statusId.toString() + "::" + getServerConfig().getNodename() + "::" + version;
         } catch (Exception e) {
@@ -297,7 +326,7 @@ public class EhrServiceImp extends BaseService implements EhrService {
             versionedEhrStatus.setUid(new HierObjectId(ehrStatus.get().getUid().toString()));
             versionedEhrStatus.setOwnerId(new ObjectRef<>(new HierObjectId(ehrUid.toString()), "local", "EHR"));
             I_EhrAccess ehrAccess = I_EhrAccess.retrieveInstance(getDataAccess(), ehrUid);
-            versionedEhrStatus.setTimeCreated(new DvDateTime(OffsetDateTime.of(ehrAccess.getInitialTimeOfVersionedEhrStatus().toLocalDateTime(),
+            versionedEhrStatus.setTimeCreated(new DvDateTime(OffsetDateTime.of(ehrAccess.getStatusAccess().getInitialTimeOfVersionedEhrStatus().toLocalDateTime(),
                     OffsetDateTime.now().getOffset())));
         }
 
@@ -309,7 +338,7 @@ public class EhrServiceImp extends BaseService implements EhrService {
         I_EhrAccess ehrAccess = I_EhrAccess.retrieveInstance(getDataAccess(), ehrUid);
 
         // get number of versions
-        int versions = ehrAccess.getNumberOfEhrStatusVersions();
+        int versions = I_StatusAccess.getLatestVersionNumber(getDataAccess(), ehrAccess.getStatusId());
         // fetch each version
         UUID versionedObjectUid = getEhrStatusVersionedObjectUidByEhr(ehrUid);
         RevisionHistory revisionHistory = new RevisionHistory();