+++ CHANGELOG.md
@@ -9,11 +9,25 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-
 ### Changed
 
+### Fixed
+
+## [0.13.0] (beta)
+
+### Added
+- added support for various functions in AQL (aggregation, statistical, string etc.) (see: https://github.com/ehrbase/ehrbase/pull/223/)
+
+### Changed
+- Update of AQL-Query test suite (see: https://github.com/ehrbase/ehrbase/pull/179)
 
 ### Fixed
+- Added validation checking for other_details and ehr_status. (see: https://github.com/ehrbase/ehrbase/pull/207)
+- Supports archetype_node_id and name for EHR_STATUS (see: https://github.com/ehrbase/ehrbase/pull/207)
+- fixes bad canonical encoding for observation/data/origin (see: https://github.com/ehrbase/ehrbase/pull/213)
+- POST without accept header for ehr, composition and contribution endpoints (see: https://github.com/ehrbase/ehrbase/pull/199)
+- force a default timezone if not present for context/start_time and context/end_time if specified (https://github.com/ehrbase/ehrbase/pull/215)
+- Representation of version uid of EHR_STATUS (see: https://github.com/ehrbase/ehrbase/pull/180)
 
 
 ## [0.12.0] (alpha)
@@ -25,8 +39,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Contribution endpoint checks for some invalid input combinations (see: https://github.com/ehrbase/ehrbase/pull/202)
 - Fixes response code on /ehr PUT with invalid ID (see: https://github.com/ehrbase/project_management/issues/163)
 - Fixes STATUS w/ empty subject bug (see: https://github.com/ehrbase/ehrbase/pull/196)
+- Now querying on composition category returns the correct result (composition/category...)
 - Fixes storage of party self inside compositions (see: https://github.com/ehrbase/ehrbase/pull/195)
 - Added support of AQL query in the form of c/composer (see: https://github.com/ehrbase/ehrbase/pull/184)
 - Java error with UTF-8 encoding resolved (see: https://github.com/ehrbase/ehrbase/pull/173)
@@ -74,6 +90,7 @@ Starting with the next release this file will provide a proper overview.
 - Correct resolution of ehr/system_id in AQL (see: https://github.com/ehrbase/ehrbase/pull/102)
 - Detection of duplicate aliases in aql select (see: https://github.com/ehrbase/ehrbase/pull/98)
 
+
 ## [0.10.0] (alpha)
 
 ### Added
@@ -114,7 +131,9 @@ Starting with the next release this file will provide a proper overview.
 - Fixed error regarding DISTINCT operator in AQL (https://github.com/ehrbase/ehrbase/issues/50)
 - Fixed null pointer exceptions that could occur in persistent compositions
 
-[unreleased]: https://github.com/ehrbase/ehrbase/compare/v0.10.0...HEAD
+[unreleased]: https://github.com/ehrbase/ehrbase/compare/v0.13.0...HEAD
+[0.13.0]: https://github.com/ehrbase/ehrbase/compare/v0.12.0...v0.13.0
+[0.12.0]: https://github.com/ehrbase/ehrbase/compare/v0.11.0...v0.12.0
 [0.11.0]: https://github.com/ehrbase/ehrbase/compare/v0.10.0...v0.11.0
 [0.10.0]: https://github.com/ehrbase/ehrbase/compare/v0.9.0...v0.10.0
 [0.9.0]: https://github.com/ehrbase/ehrbase/releases/tag/v0.9.0